auth tables working in neon! google still buggy
