+++ src/server/db/schema.ts
@@ -1,7 +1,6 @@
 import { sql } from "drizzle-orm";
-import { index, pgTableCreator} from "drizzle-orm/pg-core";
+import { pgTable, pgTableCreator, index, text, timestamp, boolean } from "drizzle-orm/pg-core";
 import { type Message } from "ai"; 
-
 
 /**
  * This is an example of how to use the multi-project schema feature of Drizzle ORM. Use the same
@@ -9,7 +8,7 @@ import { type Message } from "ai";
  *
  * @see https://orm.drizzle.team/docs/goodies#multi-project-schema
  */
-export const createTable = pgTableCreator((name) => `t3-chat-app_${name}`); 
+export const createTable = pgTableCreator((name) => `tarot-chat-app_${name}`); 
 
 type Messages = Message[]
 
@@ -45,23 +44,48 @@ export const posts = createTable(
   (t) => [index("name_idx").on(t.name)],
 );
 
-/*
-export const embeddings = pgTable('embeddings', {
-    id: varchar('id', { length: 191 })
-      .primaryKey()
-      .$defaultFn(() => nanoid()),
-    resourceId: varchar('id', { length: 256 }).references(
-      () => posts.id,
-      { onDelete: 'cascade' },
-    ),
-    content: text('content').notNull(),
-    embedding: vector('embedding', { dimensions: 1536 }).notNull(),
-  },
-  table => ({
-    embeddingIndex: index('embeddingIndex').using(
-      'hnsw',
-      table.embedding.op('vector_cosine_ops'),
-    ),
-  }),
-);
-*/
+export const user = pgTable("user", {
+					id: text('id').primaryKey(),
+					name: text('name').notNull(),
+ email: text('email').notNull().unique(),
+ emailVerified: boolean('email_verified').$defaultFn(() => false).notNull(),
+ image: text('image'),
+ createdAt: timestamp('created_at').$defaultFn(() => /* @__PURE__ */ new Date()).notNull(),
+ updatedAt: timestamp('updated_at').$defaultFn(() => /* @__PURE__ */ new Date()).notNull()
+				});
+
+export const session = pgTable("session", {
+					id: text('id').primaryKey(),
+					expiresAt: timestamp('expires_at').notNull(),
+ token: text('token').notNull().unique(),
+ createdAt: timestamp('created_at').notNull(),
+ updatedAt: timestamp('updated_at').notNull(),
+ ipAddress: text('ip_address'),
+ userAgent: text('user_agent'),
+ userId: text('user_id').notNull().references(()=> user.id, { onDelete: 'cascade' })
+				});
+
+export const account = pgTable("account", {
+					id: text('id').primaryKey(),
+					accountId: text('account_id').notNull(),
+ providerId: text('provider_id').notNull(),
+ userId: text('user_id').notNull().references(()=> user.id, { onDelete: 'cascade' }),
+ accessToken: text('access_token'),
+ refreshToken: text('refresh_token'),
+ idToken: text('id_token'),
+ accessTokenExpiresAt: timestamp('access_token_expires_at'),
+ refreshTokenExpiresAt: timestamp('refresh_token_expires_at'),
+ scope: text('scope'),
+ password: text('password'),
+ createdAt: timestamp('created_at').notNull(),
+ updatedAt: timestamp('updated_at').notNull()
+				});
+
+export const verification = pgTable("verification", {
+					id: text('id').primaryKey(),
+					identifier: text('identifier').notNull(),
+ value: text('value').notNull(),
+ expiresAt: timestamp('expires_at').notNull(),
+ createdAt: timestamp('created_at').$defaultFn(() => /* @__PURE__ */ new Date()),
+ updatedAt: timestamp('updated_at').$defaultFn(() => /* @__PURE__ */ new Date())
+				});
